refactor(advanced): FSD 아키텍처를 위한 공용 모듈 분리
- 공용 훅스들을 shared/hooks로 이동
- 유틸리티 함수들을 shared/lib로 이동
