+++ src/advanced/hooks/useCouponForm.ts
@@ -2,8 +2,8 @@ import { useState } from 'react';
 
 import type { Coupon } from '../../types';
 import { initialCouponForm, type NotificationVariant } from '../constants';
-import { useForm } from '../utils/hooks/useForm';
-import { isValidCouponCode } from '../utils/validators';
+import { useForm } from '../shared/hooks';
+import { isValidCouponCode } from '../shared/lib';
 
 interface UseCouponFormProps {
   onAddCoupon: (coupon: Coupon) => void;
