+++ src/advanced/hooks/useCouponStore.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react';
 
 import type { Coupon } from '../../types';
 import { initialCoupons } from '../constants';
-import { useLocalStorage } from '../utils/hooks/useLocalStorage';
+import { useLocalStorage } from '../shared/hooks';
 
 export function useCouponStore() {
   const [coupons, setCoupons] = useLocalStorage<Coupon[]>('coupons', initialCoupons);
