+++ src/advanced/shared/hooks/index.ts
@@ -1,0 +1,3 @@
+export { useDebounce } from './useDebounce';
+export { useForm } from './useForm';
+export { useLocalStorage } from './useLocalStorage';
